feat: Add binding to toggle sandbar location
I prefer the bar at the top, but have noticed that it can feel really
busy when you have another “toolbar” sitting right underneath it.

I’m unsure whether this is going to be the solution I’ll settle on, or
whether I’ll try to convince Mikey to add an option to move *his*
toolbar to the bottom of the interface in Zephyr controller.