feat: Convert build to use Sphinx
I’m pretty happy with river at this point, so it looks like it will be sticking
around. The biggest change that means is that I would like to start referencing
the river configuration from within my other documentation sets, and switching
to Sphinx makes that easy.

Thanks Sphinx people!