chore: Update river to 0.3.13
Recent tags are only available from the river-classic repository, so
we’ll add a note about switching to that.

This is actually the first I’d heard of the change, and it looks great!
Breaking up the project in the planned manner will make it far easier to
have custom layout engines, perhaps even something more closely
resembling a awesomewm replacement.