ci: Return success for non-matching runs
The current implementation has a problem in that you can't depend on the
checks for merge rules.  If a rule doesn't need to run then it is
correctly skipped, but then the rules will block merge.

This change handles the case where a workflow doesn't need to run, but
is still required to return success for merge rules to be effective.

I assume there is a better way to handle this, but I can't for the life
of me figure it out.